+++ packages/form/src/components/List/ListContainer.tsx
@@ -1,8 +1,9 @@
 ﻿import { PlusOutlined } from '@ant-design/icons';
 import { useIntl } from '@ant-design/pro-provider';
-import { nanoid, runFunction } from '@ant-design/pro-utils';
+import { nanoid, runFunction, useToken } from '@ant-design/pro-utils';
 import { Button } from 'antd';
 import omit from 'omit.js';
+import type { CSSProperties } from 'react';
 import { useMemo, useRef, useState } from 'react';
 import type { ProFormListItemProps } from './ListItem';
 import { ProFormListItem } from './ListItem';
@@ -20,7 +21,12 @@ const ProFormListContainer: React.FC<ProFormListItemProps> = (props) => {
     max,
     fieldExtraRender,
     meta,
+    containerClassName,
+    containerStyle,
+    onAfterAdd,
+    onAfterRemove,
   } = props;
+  const { hashId } = useToken();
   const fieldKeyMap = useRef(new Map<string, string>());
   const [loading, setLoading] = useState(false);
 
@@ -48,9 +54,17 @@ const ProFormListContainer: React.FC<ProFormListItemProps> = (props) => {
       wrapAction.add = async (...rest) => {
         const success = await actionGuard.beforeAddRow!(...rest, count);
         if (success) {
-          return action.add(...rest);
+          const res = action.add(...rest);
+          onAfterAdd?.(...rest, count + 1);
+          return res;
         }
         return false;
+      };
+    } else {
+      wrapAction.add = async (...rest) => {
+        const res = action.add(...rest);
+        onAfterAdd?.(...rest, count + 1);
+        return res;
       };
     }
 
@@ -58,14 +72,29 @@ const ProFormListContainer: React.FC<ProFormListItemProps> = (props) => {
       wrapAction.remove = async (...rest) => {
         const success = await actionGuard.beforeRemoveRow!(...rest, count);
         if (success) {
-          return action.remove(...rest);
+          const res = action.remove(...rest);
+          onAfterRemove?.(...rest, count - 1);
+          return res;
         }
         return false;
+      };
+    } else {
+      wrapAction.remove = async (...rest) => {
+        const res = action.remove(...rest);
+        onAfterRemove?.(...rest, count - 1);
+        return res;
       };
     }
 
     return wrapAction;
-  }, [action, actionGuard, uuidFields]);
+  }, [
+    action,
+    actionGuard?.beforeAddRow,
+    actionGuard?.beforeRemoveRow,
+    onAfterAdd,
+    onAfterRemove,
+    uuidFields.length,
+  ]);
 
   const creatorButton = useMemo(() => {
     if (creatorButtonProps === false || uuidFields.length === max) return null;
@@ -75,7 +104,7 @@ const ProFormListContainer: React.FC<ProFormListItemProps> = (props) => {
     } = creatorButtonProps || {};
     return (
       <Button
-        className={`${prefixCls}-creator-button-${position}`}
+        className={`${prefixCls}-creator-button-${position} ${hashId}`}
         type="dashed"
         loading={loading}
         block
@@ -100,19 +129,21 @@ const ProFormListContainer: React.FC<ProFormListItemProps> = (props) => {
     max,
     intl,
     prefixCls,
+    hashId,
     loading,
     wrapperAction,
     creatorRecord,
   ]);
 
+  const defaultStyle: CSSProperties = {
+    width: 'max-content',
+    maxWidth: '100%',
+    minWidth: '100%',
+    ...containerStyle,
+  };
+
   return (
-    <div
-      style={{
-        width: 'max-content',
-        maxWidth: '100%',
-        minWidth: '100%',
-      }}
-    >
+    <div style={defaultStyle} className={containerClassName}>
       {creatorButtonProps !== false && creatorButtonProps?.position === 'top' && creatorButton}
       {uuidFields.map((field, index) => {
         return (